+++ mmpose/datasets/pipelines/bottom_up_transform.py
@@ -117,9 +117,9 @@ def _resize_align_multi_scale_udp(image, input_size, current_scale, min_scale):
 
     trans = get_warp_matrix(
         theta=0,
-        size_input=np.array(scale, dtype=np.float),
-        size_dst=np.array(size_resized, dtype=np.float) - 1.0,
-        size_target=np.array(scale, dtype=np.float))
+        size_input=np.array(scale, dtype=np.float32),
+        size_dst=np.array(size_resized, dtype=np.float32) - 1.0,
+        size_target=np.array(scale, dtype=np.float32))
     image_resized = cv2.warpAffine(
         image.copy(), trans, size_resized, flags=cv2.INTER_LINEAR)
 
@@ -457,8 +457,8 @@ class BottomUpRandomAffine:
                     theta=aug_rot,
                     size_input=center * 2.0,
                     size_dst=np.array(
-                        (_output_size, _output_size), dtype=np.float) - 1.0,
-                    size_target=np.array((scale, scale), dtype=np.float))
+                        (_output_size, _output_size), dtype=np.float32) - 1.0,
+                    size_target=np.array((scale, scale), dtype=np.float32))
                 mask[i] = cv2.warpAffine(
                     (mask[i] * 255).astype(np.uint8),
                     trans, (int(_output_size), int(_output_size)),
@@ -472,8 +472,9 @@ class BottomUpRandomAffine:
                 theta=aug_rot,
                 size_input=center * 2.0,
                 size_dst=np.array(
-                    (self.input_size, self.input_size), dtype=np.float) - 1.0,
-                size_target=np.array((scale, scale), dtype=np.float))
+                    (self.input_size, self.input_size), dtype=np.float32) -
+                1.0,
+                size_target=np.array((scale, scale), dtype=np.float32))
             image = cv2.warpAffine(
                 image,
                 mat_input, (int(self.input_size), int(self.input_size)),
